pyproject.toml
[project]
name = "thermodynamics"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "chardet>=5.2.0",
    "python-dotenv>=1.1.0",
]
